+++ CHANGELOG.md
@@ -10,6 +10,12 @@ All notable changes to this project will be documented in this file.
   of `In` and `Out`
 
 
+### Added
+
+- disco: add `Hash` and `AppendHash` methods to the `Info` type for calculating
+  entity capabilities hashes
+
+
 ## v0.20.0 — 2021-09-26
 
 ### Security